pyproject.toml
[project]
name = "infernis"
version = "0.1.0"
description = "INFERNIS - BC Forest Fire Prediction Engine"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "xgboost>=2.0",
    "scikit-learn>=1.4",
    "pandas>=2.2",
    "geopandas>=1.0",
    "numpy>=1.26",
    "rasterio>=1.3",
    "shapely>=2.0",
    "pyproj>=3.6",
    "cdsapi>=0.7",
    "earthengine-api>=1.1",
    "sqlalchemy>=2.0",
    "geoalchemy2>=0.15",
    "psycopg2-binary>=2.9",
    "alembic>=1.13",
    "redis>=5.0",
    "apscheduler>=3.10",
    "httpx>=0.27",
    "joblib>=1.3",
    "netCDF4>=1.7",
    "xarray>=2024.1",
    "scipy>=1.12",
    "python-dotenv>=1.0",
    "pyarrow>=14.0",
    "Pillow>=10.0",
    "torch>=2.0",
    "sentry-sdk[fastapi]>=2.0",
    "cfgrib>=0.9.10",
]

[project.optional-dependencies]
dashboard = [
    "firebase-admin>=6.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "ruff>=0.5",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
